Feat: Extract page logic into separate classes and update imports (#8)
